pyproject.toml
[project]
name = "smart-calendar-agent"
version = "0.1.0"
description = "Unified calendar agent that ingests events, dedupes, and suggests self-care slots."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.1",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.1",
  "google-api-python-client>=2.130.0",
  "google-auth>=2.29.0",
  "google-auth-oauthlib>=1.2.0",
  "icalendar>=5.0.12",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.2.2",
  "dateparser>=1.2.0",
  "python-dateutil>=2.9.0",
  "openai>=1.30.5",
  "tenacity>=8.3.0",
  "email-normalize>=2.0.0",
  "pyyaml>=6.0.1",
  "cachetools>=5.3.3",
  "rapidfuzz>=3.8.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.1",
  "pytest-asyncio>=0.23.7",
  "pytest-cov>=5.0.0",
  "ruff>=0.4.8",
  "mypy>=1.10.0",
  "types-python-dateutil",
  "types-requests",
  "httpx[socks]"
]

[build-system]
requires = ["setuptools>=65.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "core", "services"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
